feat: add time utils
