+++ internal/common/utils/time_utils.go
@@ -1,0 +1,18 @@
+// Copyright 2022 Tatris Project Authors. Licensed under Apache-2.0.
+
+package utils
+
+import "time"
+
+func Timestamp2Unix(n int64) time.Time {
+	if n > 1e18 {
+		return time.Unix(0, n)
+	}
+	if n > 1e15 {
+		return time.UnixMicro(n)
+	}
+	if n > 1e12 {
+		return time.UnixMilli(n)
+	}
+	return time.Unix(n, 0)
+}
