+++ internal/common/utils/timer_utils.go
@@ -24,16 +24,3 @@ func Timerf(format string, a ...any) func() {
 	msg := fmt.Sprintf(format, a...)
 	return Timer(msg)
 }
-
-func Timestamp2Unix(n int64) time.Time {
-	if n > 1e18 {
-		return time.Unix(0, n)
-	}
-	if n > 1e15 {
-		return time.UnixMicro(n)
-	}
-	if n > 1e12 {
-		return time.UnixMilli(n)
-	}
-	return time.Unix(n, 0)
-}
